refactor(065): move InitialPuzzleTest into data layer
